Condense int and string elements of static-array properties.
Instead of writing nested full property objects with identical names
and kinds, just write a raw array of integers or strings. This greatly
shortens the size of these elements, especially in the barracks.

Note that for strings only arrays consisting entirely of narrow
strings are condensed. For non-INT games this might need to be looked
at again if it's common that all the elements are wide strings, as they
won't get codensed. However, the only real instance of this I've seen
in the game saves is the array of medal names so it's not as big of
an issue as the int arrays in soldiers.