Bugfix for enum arrays : they can have non-zero number values.
Refactored the enum properties and enum array properties to hold
enum_value objects, which are a pair of name and integer number.

This also required rewriting the array kind heuristic to handle
reading non-zero values after a string. This could be an enum
value with a non-zero number or the next string in an array of
strings. We still cannot handle embedded empty strings (it's
ambiguous).