Add support for Android saves
Android save support now working.

Android saves use version 0x13 for the save file (vs 0x10 for the PC
version of EW). The Android saves differ from PC in several key ways:

- Android saves have no header checksum, just the compressed data
checksum. It occurs in the same place as for PC, and is the same CRC32B
algorithm.

- The compressed data section of Android saves are compressed using zlib
instead of lzo.

- There is also a small bit of data after the checksum in the header but
before the compressed data. This may be Android profile related?