Better output file handling
When -o is not specified in json2xom, derive the name of the output file
from the name of the input file by doing the following:

- Trim the filename at the last occurrence of ".json", if any.
- If the remaining filename already exists in the working directory,
append ".out"

I.e. save100.json becomes either 'save100' or 'save100.out' depending on
whether there is already a 'save100' in the working directory.