Handle ambiguity in array kind determination
Turns out that reading a "None" as the first element of an array
doesn't necessarily mean it's a struct array where the first element
has all default values -- "None" can also appear in enum arrays,
presumably where "None" means the 0 value of the enum (unconfirmed).

This has been observed in the
XGExaltSimulation.m_arrCellLastVisibilityData checkpoint element,
where the first element is "None" and the second is an enumerator
value.

Since this is ambiguous, we need to keep scanning ahead at subsequent
array elements to figure out what kind of array this really is. It's
in theory possible that all the elements will be "None", although I
haven't seen this yet. If that can happen we have no recourse except
to hard-code particular known arrays with the correct type as per the
UPK file or actually read the UPKs to determine what shape the array
has.