Fix writing of object properties in EU
The object property structure has a different size in EU and EW. When
converting from json to binary we need to know the target version in
order to produce the right format.

This adds the version to all the functions that build xcom objects from
json, although several don't need the version yet.

Fixes #20